feat: Let backend refresh itself

+++ src/client/network/supabase/index.ts
@@ -2,6 +2,7 @@ import { RealtimeChannel, createClient } from "@supabase/supabase-js";
 import { Database } from "../../../@types/supabase";
 import Channel, {
   DmChannel,
+  GroupChannel,
   InvitedChannelListing,
   PrivacyLevel,
   PublicChannelListing,
@@ -273,8 +274,17 @@ class SupabaseBackendImpl implements NetworkBackend {
   }
 
   async deleteChannel(id: number): Promise<void> {
+    const user = this.loggedInUserSubscribable.getSnapshot();
+    if (user === null || user instanceof Error)
+      throw new Error("Loading or signing in.");
     const { error } = await this.client.from("channels").delete().eq("id", id);
     if (error) throw error;
+    const currentChannelList = await this.cache.getChannelListOrFallback(() =>
+      getChannels(this.client, user.id as string)
+    );
+    this.cache.putChannelList(
+      currentChannelList.getSnapshot().filter((channel) => channel.id !== id)
+    );
   }
 
   async createChannel(
@@ -309,12 +319,20 @@ class SupabaseBackendImpl implements NetworkBackend {
     const channel = {
       ...dbChannel,
       users: [
-        await this.cache.getUserOrFallback(userId, () =>
-          getUser(this.client, userId)
-        ),
+        (
+          await this.cache.getUserOrFallback(userId, () =>
+            getUser(this.client, userId)
+          )
+        ).getSnapshot(),
       ],
     };
     this.cache.putChannel(channel);
+    const currentChannelList = await this.cache.getChannelListOrFallback(() =>
+      getChannels(this.client, user.id as string)
+    );
+    this.cache.putChannelList(
+      currentChannelList.getSnapshot().concat([channel])
+    );
     return convertChannel(channel);
   }
 
@@ -347,12 +365,13 @@ class SupabaseBackendImpl implements NetworkBackend {
   }
 
   getUser(id: string): Subscribable<User | null> {
-    return new Subscribable<User | null>(async (next) => {
-      next(
-        convertUser(
-          await this.cache.getUserOrFallback(id, () => getUser(this.client, id))
-        )
-      );
+    return new Subscribable<User | null>(async (next, nextError) => {
+      (await this.cache.getUserOrFallback(id, () => getUser(this.client, id)))
+        .map<User | null>((user) => Promise.resolve(convertUser(user)), null)
+        .subscribe(({ value, error }) => {
+          if (error) nextError(error);
+          else next(value);
+        });
     }, null);
   }
 
@@ -363,21 +382,31 @@ class SupabaseBackendImpl implements NetworkBackend {
   }
 
   getDMs(): Subscribable<DmChannel[] | null> {
-    return this.loggedInUserSubscribable.map<DmChannel[] | null>(
-      async (value) => {
-        if (!value) return null;
-        const channels = await getChannels(
-          this.client,
-          value.id as string,
-          false
-        );
-        this.cache.putChannel(...channels);
-        return channels
-          .map(convertChannel)
-          .filter((channel) => channel.dm) as DmChannel[];
-      },
-      null
-    );
+    return new Subscribable<DmChannel[] | null>(async (next, nextError) => {
+      (
+        await this.cache.getChannelListOrFallback(() => {
+          const id = this.getCurrentSession().getSnapshot()?.id;
+          return id
+            ? getChannels(this.client, id as string)
+            : Promise.resolve([]);
+        })
+      )
+        .map<DmChannel[] | null>(
+          (channels) =>
+            Promise.resolve(
+              channels
+                .map((channel) => convertChannel(channel))
+                .filter<DmChannel>(
+                  (channel): channel is DmChannel => channel.dm
+                )
+            ),
+          null
+        )
+        .subscribe(({ value, error }) => {
+          if (error) nextError(error);
+          else next(value);
+        });
+    }, null);
   }
 
   joinChannel<JoinInfo extends ChannelJoinInfo>(
@@ -387,18 +416,32 @@ class SupabaseBackendImpl implements NetworkBackend {
   }
 
   getChannels(): Subscribable<Channel[] | null> {
-    return this.loggedInUserSubscribable.map<Channel[] | null>(
-      async (value) => {
-        if (!value) return null;
-        const channels = await getChannels(this.client, value.id as string);
-        this.cache.putChannel(...channels);
-        channels.forEach((channel) => {
-          this.cache.putUser(...channel.users);
+    return new Subscribable<Channel[] | null>(async (next, nextError) => {
+      (
+        await this.cache.getChannelListOrFallback(() => {
+          const id = this.getCurrentSession().getSnapshot()?.id;
+          return id
+            ? getChannels(this.client, id as string)
+            : Promise.resolve([]);
+        })
+      )
+        .map<GroupChannel[] | null>(
+          (channels) =>
+            Promise.resolve(
+              channels
+                .map((channel) => convertChannel(channel))
+                .filter<GroupChannel>(
+                  (channel): channel is GroupChannel => !channel.dm
+                )
+            ),
+          null
+        )
+        .subscribe(({ value, error }) => {
+          console.log(value, error);
+          if (error) nextError(error);
+          else next(value);
         });
-        return channels.map(convertChannel);
-      },
-      []
-    );
+    }, null);
   }
 
   clearCache(): Promise<void> {
@@ -424,18 +467,20 @@ class SupabaseBackendImpl implements NetworkBackend {
   }
 
   getChannel(id: number): Subscribable<Channel | null> {
-    return new Subscribable<Channel | null>(async (next) => {
-      try {
-        next(
-          convertChannel(
-            await this.cache.getChannelOrFallback(id, () =>
-              getChannel(this.client, id)
-            )
-          )
-        );
-      } catch {
-        next(null);
-      }
+    return new Subscribable<Channel | null>(async (next, nextError) => {
+      (
+        await this.cache.getChannelOrFallback(id, () =>
+          getChannel(this.client, id)
+        )
+      )
+        .map<Channel | null>(
+          (user) => Promise.resolve(convertChannel(user)),
+          null
+        )
+        .subscribe(({ value, error }) => {
+          if (error) nextError(error);
+          else next(value);
+        });
     }, null);
   }
